requirements.txt
python-telegram-bot==20.3
woocommerce==3.0.0
requests
